Create code.py
In this game we need to take the user’s choice and then we need to compare it with the computer choice which is taken using the random module in Python from a list of choices, and if the user wins then the score will increase by 1. And if computer wins the score will inrease by 1.